+++ CHANGELOG.md
@@ -18,8 +18,9 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add option to specify a custom backups path
 - Added Tokyonight theme
 - Censor IP addresses from Minecraft in launcher logs [#964]
-- Add PaperMC to loader list for creating plugin capable servers [#975]
+- Add Paper to loader list for creating plugin capable servers [#975]
 - Add option to install/manage plugins and mods to servers [#975][#421]
+- Add Purpur support for servers
 
 ### Fixes
 - Invalid Java install location causing issues starting the launcher